+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 * Add `retrieval_reference_number` to `Transaction`
+* Fix class alias issue where Exceptions could not be thrown for missing libraries or older PHP versions.
 
 ## 5.1.0
 * Add `threeDSecurePassThru` parameters to `Customer.create`, `PaymentMethod.create`, `CreditCard.create`, `Customer.update`, `PaymentMethod.update` and `CreditCard.update`