+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 5.2.0
 * Add `retrieval_reference_number` to `Transaction`
 * Fix class alias issue where Exceptions could not be thrown for missing libraries or older PHP versions.
 * Fix issue where `proxyHost`, `proxyPort`, `proxyType`, `proxyUser`, `proxyPassword`, and `sslVersion` were not being set on Gateway or Configuration objects.